services/user-manager/backend/requirements.txt
uvicorn[standard]
fastapi 
uvicorn
mysqlclient
sqlalchemy==2.0.10
pydantic==1.10.7
python-jose
passlib
